IMPALA-5426: Update Hive schema script to 1.1.0
A recent update to Hive changed its schema, which is causing
the metastore not to come up when run against the latest version
of Hive as we've hard coded the Hive schema script version in
bin/create-test-configuration.sh to an old version.

This patch updates the version to the latest. The schema script
is included in Hive in the toolchain and the new version will
already be present.

By itself, this patch does not actually change the Hive schema
by default as the version we usually build against doesn't have
the change. A following patch will update impala-config.sh to pull
in the latest version of Hive.

In the long run, we should switch to using Hive's schema tool,
which can do this for us automatically (IMPALA-5430).

Testing:
- Ran an exhaustive private Jenkins build that passed.

Change-Id: I9ea3269c1f95f76d8c02b76a5dea3ca3aa324b70
Reviewed-on: http://gerrit.cloudera.org:8080/7072
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23456
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>