Revert "IMPALA-5428: update external hadoop ecosystem versions"
The update to hadoop versions in impala-config.sh was committed
to asf-master and automatically cherry-picked onto cdh5-trunk.

Unfortunately, the Jenkins job that updates the contents of
'/thirdparty' expects that impala-config.sh hasn't been updated
yet.

Rather than fix the job, we've been asked by Releng to simply
revert the changes to impala-config.sh so that the job can pass.

This reverts commit e8acc8157bc2495ec80087bb9d01ec99500d9764,
except for the update to KUDU_JAVA_VERSION, which is not handled
by the job because it is pulled in through maven rather than
contained in '/thirdparty'

Change-Id: If3ce829e42b443e34d762712ccdda916598e5fbe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23485
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>