IMPALA-5469: Fix exception when processing catalog update
In IMPALA-5355 patch, a regression was introduced where we end up
calling addCatalogObject on a CATALOG object type in the second pass
over the catalog objects because the CATALOG object is not considered
a top level object.

The issue is fixed by updating the top level objects in the first pass
and updating the non top level objects and the catalog object in the
second pass.

Testing:
- Verified by manually running "invalidate metadata" and checking that
  there is no exception in the impalad.INFO log file.
- No new tests were added because it is difficult and excessive to
  reproduce the issue in a test.

Change-Id: I7b8e7db2c679ea9bffd8863a3367aa3b94e4e1ed
Reviewed-on: http://gerrit.cloudera.org:8080/7121
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 87a7a6418eacea9e1be6757baa40b6df3f72f4dd)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23872
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>