IMPALA-5536: Fix TCLIService thrift compilation on Hive 2
This fixes some issues with Impala compilation on Hive 2:

1. It fixes TCLIService.thrift generation by changing
common/thrift/CMakeLists.txt to create the directory
before generating the file and adding a dependency
to generate TCLIService.thrift before compiling the
thrift files.
2. It modifies CatalogOpExecutor::applyAlterTable to
use the MetastoreShim.

With these changes, Impala builds successfully with
Hive 2 in my environment and on automated builds.

Change-Id: I7e1477993ee3ccddd236609efec7bb23f20a7b66
Reviewed-on: http://gerrit.cloudera.org:8080/7231
Tested-by: Impala Public Jenkins
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>