IMPALA-4687: Get Impala working against HBase 2.0
This changes Impala code to tolerate the API differences
between HBase 1.0 and HBase 2.0. It also drops
compatibility code for older HBase versions.

Specific changes:
1. Tolerate return value of Scan for Scan.setCaching()
and Scan.setCacheBlocks(). This has no impact on our code.
2. HBase 2.0 eliminates the ScannerTimeoutException. The
case that previously generated the exception will now
recreate the scanner, so it is not necessary for our code
to recreate the scanner. Short-circuit
HandleResultScannerTimeout on HBase 2.0.
3. HBase 2.0 eliminates the Put.add(), which has been
replaced with Put.addColumn(). This API exists in
HBase 1.0, so it is safe to switch this completely.

This was tested by verifying that an HBase 2.0 cluster
starts up.

Change-Id: I87610e25c01b3547ec332c6975b61284b6837d27
Reviewed-on: http://gerrit.cloudera.org:8080/7277
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Impala Public Jenkins