IMPALA-5650: Make sum_init_zero a SUM function
The recent Parquet count(*) optimization (IMPALA-5036) introduced a new
aggregate function sum_init_zero(). This caused an issue for non
partitioned aggregation node, which checks that the function type is
either count, min, max, sum or ndv. Since sum_init_zero() was not
labelled as a sum() function, a DCHECK was hit. This issue is fixed by
labelling sum_init_zero() as a sum() function.

Testing:
- Verified that we no longer hit a DCHECK when running a query with a
  Parquet count(*) optimization with legacy agg node enabled.

Change-Id: Ibca53e3ef6bad5283550100b33db55d674cff0b9
Reviewed-on: http://gerrit.cloudera.org:8080/7404
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Impala Public Jenkins