IMPALA-5686: Update a mini cluster Sentry property
The property "sentry.service.client.server.rpc-address" was recently
updated to "sentry.service.client.server.rpc-addresses", and as a
result, the mini cluster fails to start.

We fix the problem by adding the new property to the sentry-site.xml
template file. The old property is temporarily kept as a precaution.

Change-Id: I486c1ab33277f4e795be9dbbefb8d839354bf6bb
Reviewed-on: http://gerrit.cloudera.org:8080/7469
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins