IMPALA-5689: Avoid inverting non-equi left joins
When checking if a join can be inverted, we forgot to also check that
the resulting join would not be a non-equi right semi-join or a
non-equi right outer-join. We currently do not support those kinds of
joins in the backend.

Testing:
-Added a planner test

Change-Id: I91ba66fe30139fcd44d4615a142f183266800aab
Reviewed-on: http://gerrit.cloudera.org:8080/7476
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 0a99105856610cc6d498ea3835e6e6a9060543e3)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26569
Reviewed-by: Attila Jeges <attilaj@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>