IMPALA-5716: Don't delete cmake_modules/* when enabling distcc
If $IMPALA_HOME ends with a /, the clean_cmake_files function in
distcc_env.sh will emit a find command with a double // at the end for
the cmake_modules directory, and since it contains the substring cmake,
find will match and delete its contents.

Fix is to use a whitelist of locations and filenames to look for, and
delete only those.

Testing: manually ran enable_distcc, observed that my files were still
there.

Change-Id: I8a6e34bedf8000aed9e2b0597cfe86f73222c6ed
Reviewed-on: http://gerrit.cloudera.org:8080/7493
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins