IMPALA-5691: recalibrate mem limit for Q18
I ran the stress test binary search locally and it produced a slightly
higher number for Q18 than the hardcoded value. This is enough to move
it above one of the thresholds, so may reduce flakiness.

Testing:
I wasn't able to reproduce the flakiness locally, so can't confirm
this fixes it.

Change-Id: I1ffa969061a52730c5147d142dcd2e3cb3626590
Reviewed-on: http://gerrit.cloudera.org:8080/7512
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins