IMPALA-1891: Statestore won't send deletions in initial non-delta topic
Currently, when a subscriber connects to the statestore, its first
update contains the entire topic including deletions which are useless
because the subscriber has no base to apply those deletions to. This
patch ensures that the deletions are not included in that first update.

Testing:
Modified existing test case to include this check

Change-Id: I3fc525e1f3d960d642fc6356abb75f744cab7c33
Reviewed-on: http://gerrit.cloudera.org:8080/7527
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins