IMPALA-1470: Fix error message with catalog down
When the catalog server cannot be reached, currently impala client would report
a NullpointerException to shell with no context. The improved error message
with this patch is:
> ERROR: InternalException: Error requesting prioritized load:
Couldn't open transport for localhost:26000 (connect() failed: Connection refused)
> Error making an RPC call to Catalog server.

The bug is when the client gets an RPC error, it tries to box the error into a
field of a Thrift struct but does not set the optional field to be present.
This patch makes the status field in TPrioritizeLoadResponse required.

Change-Id: Ic695faf89b146bdaa145aaca87f85579137b67c5
Reviewed-on: http://gerrit.cloudera.org:8080/7539
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins