IMPALA-4737: Prevent SIGUSR1 from killing daemons when minidumps are disabled
If a user disabled minidumps before this change, we did not register the
signal handler for SIGUSR1 at all. Sending SIGUSR1 to a daemon would
subsequently kill it.

This change registers the SIG_IGN handler to ignore the signal if
minidumps are disabled.

Change-Id: I13d866a2eec832500131954a7f693c33585ea51e
Reviewed-on: http://gerrit.cloudera.org:8080/7631
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 400454277d865e32372b4ed25befb38825751f45)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26571
Reviewed-by: Attila Jeges <attilaj@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>