IMPALA-5787: Dropped status in KuduTableSink::Send()
KuduTableSink drops the Status returned from WriteKuduValue(),
potentially hiding failures.

There are only two possible ways WriteKuduValue() can fail: one of the
KuduPartialRow::Set<Type> functions fails, which can only happen if
the specified type is incorrect, which would be a bug in planning and
is therefore DCHECK-able, or if TimestampValue::UtcToUnixTimeMicros()
fails, which WriteKuduValue() DCHECKs against but also has a path to
return the error in release builds.

So similarly, this patch adds a DCHECK that the status is OK, and a
path to return the error in release builds.

It also adds WARN_UNUSED_RESULT to WriteKuduValue() to prevent similar
bugs in the future.

Testing:
- Ran the Kudu tests.

Change-Id: Ia228d98ca90de8638b0afcc242c788e225fb450f
Reviewed-on: http://gerrit.cloudera.org:8080/7667
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins