Propagate HAVE_PIPE2 compile time value to files that use it
The HAVE_PIPE2 is a variable that tracks whether a platform has the
system function pipe2() present.

This value was not propagated to the appropriate file that uses it,
causing its value to always be 0, and the wrong branch to be taken
at compile time.

This fixes it by propagating the value to the file.

Change-Id: I6cdc343da35a34be8d95fbea3543d080dbc1ec29
Reviewed-on: http://gerrit.cloudera.org:8080/7705
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Impala Public Jenkins