IMPALA-5819: DCHECK in HdfsTextScanner::Close()
The DCHECK is only valid if the scanner successfully processes
the entire scan range. If the scanner is cancelled or hits an
error in-between buffers, boundary_column_ may be non-empty.

The fix is to move the DCHECK to the end of FinishScanRange(),
where it only executes if the scan range is fully processed.

Testing:
Ran TestCancellation in a loop locally for an hour.

Change-Id: I3d63f0b0c1ac82103a60b68b244156b9fbd575e0
Reviewed-on: http://gerrit.cloudera.org:8080/7756
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit ffa307eb203c409fc8e285aa3fa66d791c345388)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/28015
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>