IMPALA-5798: ASAN use-after-poison in Parquet decoder
In ParquetLevelDecoder::Init() for RLE encoding, we read the metadata
size and advance the data buffer past it. If the metadata size is
corrupted, it can cause us to incorrectly read past the end of the
buffer.

This patch checks that the metadata size is less than the total size
of the buffer, and returns an error if it isn't.

Testing:
- Ran test_scanners_fuzz.py under ASAN 500 times without hitting the
  use-after-poison (previously it would usually hit in < 100 runs).

Change-Id: I3f3d0d998f7581c7c935d98fde886f145efd61a8
Reviewed-on: http://gerrit.cloudera.org:8080/7769
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins