IMPALA-5211: Simplifying ifnull/isnull/nvl where conditional is a literal.
Added code to simplify ifnull(<literal>, <x>) to x if <literal> is NULL
or <literal> if it's not NULL. Removed a few unused Java imports.

I did the naive thing and hardcoded the list of ifnull aliases.
It looks like common/function-registry/impala_functions.py is
used to generate ScalarBuiltins.java, and a similar approach could
be taken.

There are more conditional functions where similar approaches
can be taken; IMPALA-5211 is not yet complete with just this patch.

Testing:
* Added new cases to existing unit test.
* Added a simple test case to the PlannerTest infrastructure.
* I ran the relevant test with code coverage using
  "mvn -Dtest=ExprRewriteRulesTest -Djacoco.skip=false test". Coverage
  of SimplifyConditionalsRule.java is very strong. There's one
  pre-existing "return expr" in simplifyCompoundPredicate that's
  not covered, and almost all the rest are preconditions checks.

Change-Id: I9e4b8bf6fedd595f9ea54ffb30fc71a058c7f16c
Reviewed-on: http://gerrit.cloudera.org:8080/7781
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins