IMPALA-4826: Fix error during a scan on repeated root schema in Parquet.
Having the repetition level set to REPEATED on the root schema
resulted a scan to fail with error when Impala tried to parse that
table.

As a solution, the 'REPEATED' repetition level is ignored when the
root schema is processed. The reasoning behind is that the Parquet
format description says that the repetition level of the root schema
should not be set to REPEATED anyway, so it's safe to ignore it in
case it is set to this value for some reason.

Change-Id: I7ea84589e1d122ad9d43adde46893ec0ecc5f9c4
Reviewed-on: http://gerrit.cloudera.org:8080/7870
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 923a62968ee8fa573bbd17780cf47a5024335e8d)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26640
Reviewed-by: Attila Jeges <attilaj@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>