IMPALA-5597: Try casting targetExpr when building runtime filter plan
This patch fixes a bug that fails a precondition check when generating
runtime filter plans. The lhs and rhs or join predicate might have
different types when the eq predicate function accepts wildcard-typed
parameters. In this case in existing code the types of source and target
expr will be found mismatch and an exception will be thrown when
generating runtime filters. This patch tries to cast target expr to be
of the same type as source expr. A testcase is added to joins.test

Change-Id: I0d66673e280ce13cd3a514236c0c2ecbc50475a6
Reviewed-on: http://gerrit.cloudera.org:8080/7949
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit bf5decbc18ce46446b0cd24ee431e6f85844c9d4)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/27542
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>