IMPALA-5892: Allow reporting status independent of fragment instance
Queries can hit an error that is not specific to a
particular fragment instance. For example, QueryState::StartFInstances()
calls DescriptorTbl::Create() before any fragment instances
start. This location has no reason to report status via a
particular fragment, and there is currently no way to report
status otherwise. This leads to a query hang, because the
status is never propagated back to the coordinator.

This adds the ability to report status that is not associated
with a particular fragment instance. By reporting status,
the coordinator will now correctly abort the query in the
case of the QueryState::StartFInstances() scenario described.

Change-Id: I4cd98022f1d62a999c7c80ff5474fa8d069eb12c
Reviewed-on: http://gerrit.cloudera.org:8080/7943
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 8847e4a65091814906d3ed263be8aee0347535fe)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26276
Reviewed-by: Joseph McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joseph McDonnell <joemcdonnell@cloudera.com>