IMPALA-5812: Fix NPE when joining on empty const select
A NPE is thrown during the creation of the single node plan of a query
consisting of a cross join with a constant select that returns
an empty result set. This happens because when an empty-set plan node
is created, its tupleIds_ and tblRefIds_ are initialized with the
tuple ID of a newly create tuple that does not map to any existing
tableRefs. This causes a null pre-check to fail during the creation
of the join node when it tries to fetch the tableRef from that new
tuple Id in the empty-set node but doesn't find one.

Testing:
Added a planner test.

Change-Id: I6e425dbcb442aeeac687e103774823d3f50e6436
Reviewed-on: http://gerrit.cloudera.org:8080/7971
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 5e7430be9791a419cb4e1fa5501ab605200c3f39)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26331
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>