IMPALA-5923: Print binary ID as hex in ChildQuery::Cancel()
ChildQuery::Cancel() prints a binary ID into the log which can show up
as random characters. One fix is to print it as a hex string.

I tested this by running test_cancellation::test_cancel_insert and
making sure the ID is printed as hex.

This change also removes PrintAsHex() which was broken and unused.

Change-Id: Ie1a9516d5c03524e2585255700bb84e8a301a7ee
Reviewed-on: http://gerrit.cloudera.org:8080/8050
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins