CDH-59216: Expr::CloneIfNotExists() must handle errors safely
If Expr::CloneIfNotExists() encounters an error during
ExprContext::Clone(), it leaves the vector of ExprContext
pointers in an invalid state. This is because it resizes
the vector (introducing invalid pointers) before filling
them in with ExprContext::Clone(). This leads to a crash
in ExprContext::Close(), as these invalid pointers are
accessed.

This changes the code so that the ExprContext* vector
is expanded with push_back() as the ExprContext's are
cloned. This means that the vector will be in a valid
state if ExprContext::Clone() returns an error.

Hand testing with error injection shows that this
fixes the issue.

Change-Id: Ifda50f19fdbe3b84de222813e5de9ff5b49b78df
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/26527
Reviewed-by: Joseph McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joseph McDonnell <joemcdonnell@cloudera.com>