IMPALA-5174: Suppress kudu flags that aren't relevant to Impala
Importing Kudu's util/, security/ and rpc/ code into the
Impala code base results in Kudu's gflags showing up in the log
files and the /varz webpages, for all the gflags that belong to
Kudu files that are in some way used by Impala.

For eg: If in Impala, we include kudu/rpc/messenger.h and use one
of its classes whose method definitions live in the corresponding
messenger.cc file, all the flags DEFINEd in the messenger.cc file
will show up in the Impala logs. If we do not use any code in a .cc
file, the flags that are DEFINEd in that file do not show up (this
is probably due to the compiler not running macros in files that
have only dead code).

Since we don't want these flags to be user configurable, we don't
want them to be visible. The following commit earlier included a
gflags patch that added a DEFINE_*_hidden() macro that allows us to
do just this:
https://github.com/apache/incubator-impala/commit/d1910a39fcc50ce211b95c3552c0c90b4bc37bbd

This patch was done by running the following sed command:

find $IMPALA_HOME/be/src/kudu -type f | xargs sed -i 's/^\(DEFINE_.*\)\((.*\)/\1_hidden\2/g'

There were a couple non-related macros that also got changed:
DEFINE_validator() and DEFINE_STATIC_THREAD_LOCAL(), which were
manually changed back.

Change-Id: I499c903cde92595c4a02803ecbf98ac1d41517b4
Reviewed-on: http://gerrit.cloudera.org:8080/8074
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Impala Public Jenkins