IMPALA-4951: Fix database visibility for user with only column privilege
Currently a database is not visible to a user that only has column
level privileges for tables in that database. This patch will make
the database visible, which is the expected behavior in this case.

Testing: added a test case to verify the same.

Change-Id: Id77904876729c0223fd6ace2d5e7199bd700a33a
Reviewed-on: http://gerrit.cloudera.org:8080/8168
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 00c6c9cb314dc3d6fd5f024253b1e4cc8116e33f)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/27545
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>