IMPALA-5951: Remove flaky test_catalogd_timeout
test_catalogd_timeout sets a Kudu operation timeout of 1ms and then
performs various Kudu operations which it expects to fail due to a
timeout.

Since the test was written, things have sped up - for example, Impala
used to create a new Kudu client for each operation, but that was
changed in IMPALA-5167, such that the operations now occasionally
complete quickly enough that they don't timeout.

There's not really any way to rewrite this test to ensure that it
won't be flaky, so the patch removes it.

Change-Id: I29fd67d0acc0ee15943c416f2179ad716d2cac05
Reviewed-on: http://gerrit.cloudera.org:8080/8154
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit fc54882d9a4769263d800d0654048f0ab1252153)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/29751
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>