IMPALA-6009: Upgrade Guava to 14.0.1
Builds are breaking because the latest hive-exec
snapshot jar includes Guava 14.0.1 classes which
conflict with the version Impala depends on (11.0.2).

Between versions 11.0.2 and 14.0.1 Guava changed the
API of some Hasher methods that Impala uses.

As a workaround this patch upgrades Impala's Guava
dependency to version 14.0.1 to be consistent with
the classes in hive-exec.

Testing:
- mvn compile succeeded

Change-Id: Iddc5da8849d5aa7317d3dc572884d05dee859bdd
Reviewed-on: http://gerrit.cloudera.org:8080/8198
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Alex Behm <alex.behm@cloudera.com>