CDH-58865: Bump httpcomponents core version.
In CDH6, the CDH http core version is currently 4.4.6.  This commit
reconciles what Impala is using to that, preventing breakages in KMS.

This commit does not address the problem that we are still using an old
Thrift (CDH-40029) nor that this mechanism of choosing an httpcore
version can get stale (IMPALA-4210). It does not address using the
CDH root pom (CDH-43101).

For testing, I built a parcel with this change, deployed it on the
nightly6x cluster, and ran a "select * from customers" query, which had
been failing before.

Change-Id: Ic794f8981cecbc418024b68a58813776e7f1d3ff