CDH-59928: Fix / hack copy-udfs-udas.sh
Two problems conspire to cause these libudasample.so failures;
first, building with -notest during the data load loses the
test dependencies, so libudasample.so never gets built.

Second, even when it does get built, our change to append the
GBN to $HIVE_VERSION causes a failure to copy in the hive-exec.jar
This is coupled to the way we build the hive tarball, and is
something we'll need to be aware of in automating that process.

I'm not entirely happy hacking around this like so, but it
does unblock the tests.

Testing: Running tests/run-custom-cluster-tests.sh \
 -k test_permanent_udfs passes test_permanent_udfs as it
 now finds libudasample.so

Change-Id: I0881c287c0991f142662dad8f20f26b0c418d913
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/27066
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Tested-by: Zachary Amsden <zamsden@cloudera.com>