CDH-59438: Fix key splitting for HBase
The HBase API for getting cluster config now returns both
the master and the region server ports.  This was confusing
the HBase splitter code as it was only expecting RegionServers
to be listed.

Further, partial failures could happen if only some of the
splits get processed.  This would manifest as exceptions being
raised when attempting to re-split already split regions.
Instead, we just catch the error and continue.

A number of API changes seem to have made it in as well, which
broke compilation in a number of ways.  This will need to be
updated again to remove all deprecated HBase APIs.

Change-Id: Ibac421c0ad2a0d61958d5a3761e4a9f89b70c59b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/27328
Reviewed-by: Zachary Amsden <zamsden@cloudera.com>
Tested-by: Zachary Amsden <zamsden@cloudera.com>