IMPALA-6100: increase test_exchange_delays timeout on slow builds
The timeout is doubled on build types known to be slow to reduce the
odds of hitting the flakiness.

Testing:
I wasn't able to reproduce the original problem locally. However I was
able to reproduce a similar problem by decreasing the timeout to 2000ms,
which suggests that > 1s hiccups are possible with an ASAN build, so
increasing the timeout could help.

Change-Id: Ib00cb3c63c123b7ff263e21c7ee41fa6af12606f
Reviewed-on: http://gerrit.cloudera.org:8080/8357
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Impala Public Jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/29706
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>