CDH-61156: Catch sentry exceptions in Impala
We shouldn't have different behavior for real users and non-existing
users as this leaks information about actual user names.  Sentry
started throwing exceptions for this in C6 and we can just swallow
the exception as a workaround.

Testing: run the authorization/test_grant_revoke test; it still
doesn't pass, but fails with another error:

MESSAGE: InternalException: Error making 'grantPrivilege*' RPC

This looks to me like a Sentry problem, we'll have to follow up with
Sentry team.

Change-Id: I1ca02ca9248bd4ab1fe72a564295cd69e474c547
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/27707
Reviewed-by: Philip Zeyliger <philip@cloudera.com>
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Jenkins User
Reviewed-by: Brian Towles <btowles@cloudera.com>
Reviewed-by: Zachary Amsden <zamsden@cloudera.com>