Configuring Hive logging for log4j2.
Configs our minicluster to log Hive Metastore and HiveServer2 logs into
cluster/logs, as expected. The log4j2.properties file here is from
Hive's defaults, with only hive.log.dir updated to our preferred
location.

Hive 2.x switched to using Log4j2 for its logging by default. When you
invoke "bin/hive", it eventually invokes Java with the Hive jars
first in the classpath, so the slf4j binding to log4j2 wins, and you've
got log4j2. Log4j2 needs a hive-log4j2.properties file, which this
commit has added.

The nuance in this commit is that it was tricky to figure what
hive-log4j2.properties was being used (as it turns out, one
from one of the Hive jars) and to put something ahead of it on the
classpath. The latter was tricky because Hadoop's shell scripts
de-dupe the classpath, but you lose ordering.

Other than that, I tested this by running it and observing the logs
where you'd expect.

Change-Id: I18339099eabff1703fc822bd5ed9868fdb3f6e09