CDH-59860: Point Impala (branch cdh6.x) to CDH GBNs for miniclusters.
impala-config-branch.sh now looks up the latest CDH GBN that
has "minicluster tarballs" and configures IMPALA_<COMPONENT>_URL to
download that build. This depends on IMPALA-6148 as well
as a change to CDH to create these tarballs.

With this change, the frontend, which is allegedly using GBN-based maven
repositories (but see next paragraph), and the minicluster will always
be on the same build of CDH. Swapping between different versions of CDH
will require changing one environment variable (and the previous builds
to still be around).

In practice, the Maven stuff is not consistent right now. There
are multiple repositories with the same artifacts, and Maven
has more discretion than we'd like on which ones to choose.
There are also repos that are unused. Ultimately, this commit
is leaving that to future work.

The lifecycle of these GBNs is tricky. I chose to stick the evaluated
GBN into toolchain/cdh_components/cdh-gbn.sh. This will hopefully avoid
churn for developers, while making the world march on with the latest
builds.

The change to copy-udfs-udas.sh changes the approach taken in previous
commit 9be9e666f17a8de7d267de9f96454f2750029924 to find hive-exec.jar.
Our current source of tarballs does not have convenience, versionless
symlinks for hive-exec.jar.

Change-Id: I4628d86022d4bd8b762313f7056d76416a58b423