CDH-59307: Disable Hive operation log in minicluster
Hive's Operation Log feature is currently leaking memory during our data
load, leading data load to run out of memory sometimes. Turning it off
"fixes" this.

I tested with this change, and the data load works with a 1GB heap,
whereas it would fail consistently at that point before.

To debug this sort of thing, the following in
testdata/bin/run-hive-server.sh is the place to force Java
to dump a heap on OOM:

  HADOOP_HEAPSIZE="1024" HADOOP_CLIENT_OPTS="-Dhive.log.file=hive-server2.log -XX:+HeapDumpOnOutOfMemoryError -XX:HeapDumpPath=${LOGDIR}"

Change-Id: Idec61ecd9507718e5cb3239a2cd19523acb2bc86