IMPALA-4132: Use -fno-omit-frame-pointer
Using -fno-omit-frame-pointer would keep the frame pointers for
functions in the register. As a result we expect more useful stack
traces to be produced.

For testing performance benchmark was executed on TPC-H and TPC-DS
without any significant discrepancy from the baseline results.
(For the specific measures check the attachments in the Jira.)

Change-Id: Ib7d0d88ba015a847356ed0274fe91017b98cb941
Reviewed-on: http://gerrit.cloudera.org:8080/8612
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins