CDH-62403 restore postgresql-jdbc jar to fix job
https://master-02.jenkins.cloudera.com/job/impala-asf-master-exhaustive-integration/17/
is failing because of missing 3rd party dependency
http://repos.jenkins.cloudera.com/cdh5-static/cdh/5/hive-1.1.0-cdh5.14.0-SNAPSHOT.tar.gz
which I believe is created using the process located at
https://wiki.cloudera.com/display/engineering/CDH-CM-5.14.0
The "update Impala 3rd party" step in the checklist at
https://wiki.cloudera.com/display/engineering/CDH-CM-5.14.0 is blocked
because the Jenkins job at
https://master-02.jenkins.cloudera.com/job/impala-utility-thirdparty-branching/1/
failed. The failure was caused by a missing
Impala/thirdparty/postgresql-jdbc dir. The
impala-utility-thirdparty-branching job needs to place the
postgresql-jdbc jar file into thirdparty/sentry. It was thought that the
postgresql-jdbc jar was no longer needed, so it was removed in change
7a4d3efe4c357e4b3d3c3f17095ae07112b0f223. Indeed, it seems very awkward
to add the postgresql-jdbc jar to sentry like this and should be
improved, but for now I'm simply restoring the jar file so that we can
finish updating the Impala 3rd party packages in the new
cdh5-2.11.0_5.14.x branch.

This change should be made in cdh5-trunk so that future branches aren't
affected by this problem until a better fix is implemented, and in
cdh5-2.11.0_5.14.x and cdh5-2.11.0_5.14.0 to fix the current problems.

Testing:
https://master-02.jenkins.cloudera.com/job/impala-utility-thirdparty-branching/8/

Change-Id: I8463a209c0cb03b49805d1ce9688efadcd905359
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/28476
Reviewed-by: Matthew Mulder <mmulder@cloudera.com>
Tested-by: Matthew Mulder <mmulder@cloudera.com>